[IGNORE] Move provisioning to a dedicated package
