+++ internal/api/core/core.go
@@ -23,6 +23,7 @@ import (
 	"github.com/perses/perses/internal/api/core/middleware"
 	"github.com/perses/perses/internal/api/shared/dependency"
 	"github.com/perses/perses/internal/api/shared/migrate"
+	"github.com/perses/perses/internal/api/shared/provisioning"
 	"github.com/perses/perses/internal/api/shared/rbac"
 	"github.com/perses/perses/internal/api/shared/schemas"
 	"github.com/perses/perses/pkg/model/api/config"
@@ -62,7 +63,8 @@ func New(conf config.Config, banner string) (*app.Runner, dependency.Persistence
 	runner.WithTasks(watcher, migrateWatcher)
 	runner.WithCronTasks(time.Duration(conf.Schemas.Interval), reloader, migrateReloader)
 	if len(conf.Provisioning.Folders) > 0 {
-		runner.WithCronTasks(time.Duration(conf.Provisioning.Interval), serviceManager.GetProvisioning())
+		provisioningTask := provisioning.New(serviceManager, conf.Provisioning.Folders)
+		runner.WithCronTasks(time.Duration(conf.Provisioning.Interval), provisioningTask)
 	}
 	if conf.Security.EnableAuth {
 		rbacTask := rbac.NewCronTask(serviceManager.GetRBAC(), persesDAO)
